feat: Implement Turn-Based Orbital Mechanics Based on ADR 0002
- Replace floating-point positions with discrete integer-based orbital positions
- Implement travel time calculation using Brachistochrone model:
  travel_turns = 2 * sqrt(base_distance / acceleration)
- Add advance_planet_positions function for turn-based orbital advancement
- Update Position struct to use orbital_position (0 to orbital_period-1)
- Update orbit_radius and orbit_period to use u32 instead of f64
- Update config files to use integer values
- Add comprehensive unit tests for new functionality

Implements issue #42

+++ src/assets/config_loader.rs
@@ -7,8 +7,8 @@ use std::fs;
 #[derive(Debug, Deserialize)]
 pub struct PlanetConfig {
     pub id: String,
-    pub orbit_radius: f64,
-    pub orbit_period: f64,
+    pub orbit_radius: u32,  // Integer distance from star
+    pub orbit_period: u32,  // Turns to complete one orbit
     pub planet_type: String,  // Planet type as string, will be converted to enum
     #[serde(default)] // This allows the field to be missing without causing an error
     pub produces: Vec<String>,  // Optional override for produces
